pyproject.toml
[project]
name = "flask-mercure-sse"
version = "0.3.0"
description = "Server Sent Events (SSE) for Flask using the Mercure protocol"
authors = [
    {"name" = "Maxime Bouroumeau-Fuseau", email = "maxime.bouroumeau@gmail.com"}
]
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
dependencies = [
    "flask>=3.0.0,<4.0.0",
    "requests>=2.31.0,<3.0.0",
    "pyjwt>=2.8.0,<3.0.0",
    "gevent>=25.9.1",
]

[project.urls]
Repository = "https://github.com/hyperflask/flask-mercure-sse"

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
    "ruff>=0.12.4",
]

[tool.ruff]
line-length = 100

[build-system]
requires = ["uv_build>=0.8.11,<0.9.0"]
build-backend = "uv_build"
